Add tests for thumbnail generation
