+++ tests/e2e/specs/images/heic.spec.ts
@@ -1,0 +1,70 @@
+import { test, expect } from '../../fixtures';
+
+test.describe( 'Images', () => {
+	test.beforeAll( async ( { requestUtils } ) => {
+		await requestUtils.deleteAllMedia();
+	} );
+
+	test( 'uploads and converts an HEIC image', async ( {
+		browserName,
+		admin,
+		page,
+		editor,
+		mediaUtils,
+	} ) => {
+		test.skip(
+			browserName === 'webkit',
+			'Needs some investigation as to why image is uploaded as PNG instead of JPEG'
+		);
+
+		await admin.createNewPost();
+
+		await page.evaluate( () => {
+			window.wp.data
+				.dispatch( 'core/preferences' )
+				.set( 'media-experiments/preferences', 'imageFormat', 'jpeg' );
+			window.wp.data
+				.dispatch( 'core/preferences' )
+				.set(
+					'media-experiments/preferences',
+					'imageLibrary',
+					'browser'
+				);
+		} );
+
+		await editor.insertBlock( { name: 'core/image' } );
+
+		const imageBlock = editor.canvas.locator(
+			'role=document[name="Block: Image"i]'
+		);
+		await expect( imageBlock ).toBeVisible();
+
+		await mediaUtils.upload(
+			imageBlock.locator( 'data-testid=form-file-upload-input' ),
+			'hill-800x600.heic'
+		);
+
+		await page.waitForFunction(
+			() =>
+				window.wp.data.select( 'media-experiments/upload' ).getItems()
+					.length === 0
+		);
+
+		const settingsPanel = page
+			.getByRole( 'region', {
+				name: 'Editor settings',
+			} )
+			.getByRole( 'tabpanel', {
+				name: 'Settings',
+			} );
+
+		await expect( settingsPanel ).toHaveText( /Mime type: image\/jpeg/ );
+		await expect( settingsPanel.getByLabel( '#837776' ) ).toBeVisible();
+
+		// No exact comparison as there can be 1-2 char differences between browsers.
+		await expect( page.locator( 'css=[data-blurhash]' ) ).toHaveAttribute(
+			'data-blurhash',
+			/DIpODxF02WA_2f/
+		);
+	} );
+} );
