+++ tests/e2e/specs/images/optimizing.spec.ts
@@ -1,6 +1,6 @@
 import type { ImageFormat, ImageLibrary } from '@mexp/upload-media';
 
-import { test, expect } from '../fixtures';
+import { test, expect } from '../../fixtures';
 
 const scenarios: {
 	outputFormat: ImageFormat;
@@ -51,11 +51,7 @@ const scenarios: {
 	},
 ];
 
-test.describe( 'Image block', () => {
-	test.beforeAll( async ( { requestUtils } ) => {
-		await requestUtils.deleteAllMedia();
-	} );
-
+test.describe( 'Images', () => {
 	test.describe( 'uploads a file and allows optimizing it afterwards', () => {
 		for ( const {
 			outputFormat,
@@ -121,7 +117,8 @@ test.describe( 'Image block', () => {
 				await expect( imageBlock ).toBeVisible();
 
 				await mediaUtils.upload(
-					imageBlock.locator( 'data-testid=form-file-upload-input' )
+					imageBlock.locator( 'data-testid=form-file-upload-input' ),
+					'wordpress-logo-512x512.png'
 				);
 
 				await page.waitForFunction(
@@ -313,7 +310,8 @@ test.describe( 'Image block', () => {
 				await expect( imageBlock ).toBeVisible();
 
 				await mediaUtils.upload(
-					imageBlock.locator( 'data-testid=form-file-upload-input' )
+					imageBlock.locator( 'data-testid=form-file-upload-input' ),
+					'wordpress-logo-512x512.png'
 				);
 
 				await page.waitForFunction(
@@ -347,150 +345,5 @@ test.describe( 'Image block', () => {
 				).toHaveAttribute( 'data-blurhash', /xuj\[M\{WB00ay~qayM\{/ );
 			} );
 		}
-	} );
-
-	test( 'uploads and converts an HEIC image', async ( {
-		browserName,
-		admin,
-		page,
-		editor,
-		mediaUtils,
-	} ) => {
-		test.skip(
-			browserName === 'webkit',
-			'Needs some investigation as to why image is uploaded as PNG instead of JPEG'
-		);
-
-		await admin.createNewPost();
-
-		await page.evaluate( () => {
-			window.wp.data
-				.dispatch( 'core/preferences' )
-				.set( 'media-experiments/preferences', 'imageFormat', 'jpeg' );
-			window.wp.data
-				.dispatch( 'core/preferences' )
-				.set(
-					'media-experiments/preferences',
-					'imageLibrary',
-					'browser'
-				);
-		} );
-
-		await editor.insertBlock( { name: 'core/image' } );
-
-		const imageBlock = editor.canvas.locator(
-			'role=document[name="Block: Image"i]'
-		);
-		await expect( imageBlock ).toBeVisible();
-
-		await mediaUtils.upload(
-			imageBlock.locator( 'data-testid=form-file-upload-input' ),
-			'hill-800x600.heic'
-		);
-
-		await page.waitForFunction(
-			() =>
-				window.wp.data.select( 'media-experiments/upload' ).getItems()
-					.length === 0
-		);
-
-		const settingsPanel = page
-			.getByRole( 'region', {
-				name: 'Editor settings',
-			} )
-			.getByRole( 'tabpanel', {
-				name: 'Settings',
-			} );
-
-		await expect( settingsPanel ).toHaveText( /Mime type: image\/jpeg/ );
-		await expect( settingsPanel.getByLabel( '#837776' ) ).toBeVisible();
-
-		// No exact comparison as there can be 1-2 char differences between browsers.
-		await expect( page.locator( 'css=[data-blurhash]' ) ).toHaveAttribute(
-			'data-blurhash',
-			/DIpODxF02WA_2f/
-		);
-	} );
-
-	test( 'detects transparency in image', async ( {
-		admin,
-		page,
-		editor,
-		mediaUtils,
-		browserName,
-	} ) => {
-		test.skip(
-			browserName === 'webkit',
-			'No cross-origin isolation in Playwright WebKit builds yet, see https://github.com/microsoft/playwright/issues/14043'
-		);
-
-		await admin.createNewPost();
-
-		await page.evaluate( () => {
-			window.wp.data
-				.dispatch( 'core/preferences' )
-				.set(
-					'media-experiments/preferences',
-					'optimizeOnUpload',
-					true
-				);
-			window.wp.data
-				.dispatch( 'core/preferences' )
-				.set(
-					'media-experiments/preferences',
-					'png_outputFormat',
-					'png'
-				);
-			window.wp.data
-				.dispatch( 'core/preferences' )
-				.set(
-					'media-experiments/preferences',
-					'imageLibrary',
-					'browser'
-				);
-		} );
-
-		await editor.insertBlock( { name: 'core/image' } );
-
-		const imageBlock = editor.canvas.locator(
-			'role=document[name="Block: Image"i]'
-		);
-		await expect( imageBlock ).toBeVisible();
-
-		await mediaUtils.upload(
-			imageBlock.locator( 'data-testid=form-file-upload-input' ),
-			'github-mark.png'
-		);
-
-		await page.waitForFunction(
-			() =>
-				window.wp.data.select( 'media-experiments/upload' ).getItems()
-					.length === 0,
-			undefined,
-			{
-				timeout: 20000, // Transcoding might take longer
-			}
-		);
-
-		const settingsPanel = page
-			.getByRole( 'region', {
-				name: 'Editor settings',
-			} )
-			.getByRole( 'tabpanel', {
-				name: 'Settings',
-			} );
-
-		await expect( settingsPanel ).toHaveText(
-			new RegExp( 'Mime type: image/png' )
-		);
-		await expect( settingsPanel ).toHaveText(
-			new RegExp( 'Has transparency: yes' )
-		);
-		await expect( settingsPanel.getByLabel( '#24292f' ) ).toBeVisible();
-		// No exact comparison as there can be 1-2 char differences between browsers.
-		await expect( page.locator( 'css=[data-blurhash]' ) ).toHaveAttribute(
-			'data-blurhash',
-			/ofITW/
-		);
 	} );
 } );
