+++ tests/e2e/specs/images/transparency.spec.ts
@@ -1,0 +1,89 @@
+import { test, expect } from '../../fixtures';
+
+test.describe( 'Images', () => {
+	test.beforeAll( async ( { requestUtils } ) => {
+		await requestUtils.deleteAllMedia();
+	} );
+
+	test( 'detects transparency in image', async ( {
+		admin,
+		page,
+		editor,
+		mediaUtils,
+		browserName,
+	} ) => {
+		test.skip(
+			browserName === 'webkit',
+			'No cross-origin isolation in Playwright WebKit builds yet, see https://github.com/microsoft/playwright/issues/14043'
+		);
+
+		await admin.createNewPost();
+
+		await page.evaluate( () => {
+			window.wp.data
+				.dispatch( 'core/preferences' )
+				.set(
+					'media-experiments/preferences',
+					'optimizeOnUpload',
+					true
+				);
+			window.wp.data
+				.dispatch( 'core/preferences' )
+				.set(
+					'media-experiments/preferences',
+					'png_outputFormat',
+					'png'
+				);
+			window.wp.data
+				.dispatch( 'core/preferences' )
+				.set(
+					'media-experiments/preferences',
+					'imageLibrary',
+					'browser'
+				);
+		} );
+
+		await editor.insertBlock( { name: 'core/image' } );
+
+		const imageBlock = editor.canvas.locator(
+			'role=document[name="Block: Image"i]'
+		);
+		await expect( imageBlock ).toBeVisible();
+
+		await mediaUtils.upload(
+			imageBlock.locator( 'data-testid=form-file-upload-input' ),
+			'github-mark.png'
+		);
+
+		await page.waitForFunction(
+			() =>
+				window.wp.data.select( 'media-experiments/upload' ).getItems()
+					.length === 0,
+			undefined,
+			{
+				timeout: 20000, // Transcoding might take longer
+			}
+		);
+
+		const settingsPanel = page
+			.getByRole( 'region', {
+				name: 'Editor settings',
+			} )
+			.getByRole( 'tabpanel', {
+				name: 'Settings',
+			} );
+
+		await expect( settingsPanel ).toHaveText(
+			new RegExp( 'Mime type: image/png' )
+		);
+		await expect( settingsPanel ).toHaveText(
+			new RegExp( 'Has transparency: yes' )
+		);
+		await expect( settingsPanel.getByLabel( '#24292f' ) ).toBeVisible();
+		// No exact comparison as there can be 1-2 char differences between browsers.
+		await expect( page.locator( 'css=[data-blurhash]' ) ).toHaveAttribute(
+			'data-blurhash',
+			/ofITW/
+		);
+	} );
+} );
